UtgiftsperiodeDto og PeriodeMedBeløpDto hører hjemme i VedtakBarnetilsynDto - og ikke i beregningRequestDto
