+++ src/main/kotlin/no/nav/familie/ef/sak/vedtak/dto/VedtakBarnetilsynDto.kt
@@ -1,12 +1,13 @@
 package no.nav.familie.ef.sak.vedtak.dto
 
-import no.nav.familie.ef.sak.beregning.barnetilsyn.PeriodeMedBeløpDto
-import no.nav.familie.ef.sak.beregning.barnetilsyn.UtgiftsperiodeDto
+import no.nav.familie.ef.sak.felles.dto.Periode
 import no.nav.familie.ef.sak.infrastruktur.exception.feilHvis
 import no.nav.familie.ef.sak.vedtak.domain.Barnetilsynperiode
 import no.nav.familie.ef.sak.vedtak.domain.PeriodeMedBeløp
 import no.nav.familie.ef.sak.vedtak.domain.Vedtak
+import java.math.BigDecimal
 import java.time.YearMonth
+import java.util.UUID
 
 data class InnvilgelseBarnetilsyn(val begrunnelse: String?,
                                   val perioder: List<UtgiftsperiodeDto> = emptyList(),
@@ -18,6 +19,26 @@ data class TilleggsstønadDto(val harTilleggsstønad: Boolean,
                              val perioder: List<PeriodeMedBeløpDto> = emptyList(),
                              val begrunnelse: String?)
 
+data class PeriodeMedBeløpDto(
+        val årMånedFra: YearMonth,
+        val årMånedTil: YearMonth,
+        val beløp: BigDecimal
+)
+
+data class UtgiftsperiodeDto(
+        val årMånedFra: YearMonth,
+        val årMånedTil: YearMonth,
+        val barn: List<UUID>,
+        val utgifter: BigDecimal
+) {
+
+    fun tilPeriode(): Periode = Periode(this.årMånedFra.atDay(1), this.årMånedTil.atEndOfMonth())
+}
+
+fun List<UtgiftsperiodeDto>.tilPerioder(): List<Periode> =
+        this.map {
+            it.tilPeriode()
+        }
 
 fun UtgiftsperiodeDto.tilDomene(): Barnetilsynperiode =
         Barnetilsynperiode(datoFra = this.årMånedFra.atDay(1),
